Move AlternateNameDeleteOp to models
